+++ src/redturtle/volto/browser/sitemap.py
@@ -5,6 +5,7 @@ from plone.registry.interfaces import IRegistry
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.interfaces import IPloneSiteRoot
 from zope.component import getUtility
+
 import datetime