+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 4.1.4 (unreleased)
 ------------------
 
+- Remove Patch to avoid auto-closed tags in portal transforms: https://github.com/plone/Products.PortalTransforms/pull/43
+  [mamico]
+
 - Remove Patch for folderish types migration https://github.com/plone/plone.volto/pull/86
   [mamico]